🔨 Refactor import paths for grapher model
