+++ api/v1/__init__.py
@@ -14,10 +14,10 @@ from sqlalchemy.orm import Session
 
 from apps.backport.datasync.datasync import upload_gzip_dict
 from etl import config, paths
-from etl.grapher import model as gm
 from etl.command import main as etl_main
 from etl.db import get_engine
 from etl.files import yaml_dump
+from etl.grapher import model as gm
 from etl.helpers import read_json_schema
 from etl.metadata_export import merge_or_create_yaml, reorder_fields
 from etl.paths import SCHEMAS_DIR
