+++ etl/steps/__init__.py
@@ -38,10 +38,10 @@ from sqlalchemy.orm import Session
 
 from apps.chart_sync.admin_api import AdminAPI
 from etl import config, files, git_helpers, paths
-from etl import grapher_model as gm
 from etl.config import OWID_ENV, TLS_VERIFY
 from etl.db import get_engine
 from etl.grapher import helpers as gh
+from etl.grapher import model as gm
 from etl.snapshot import Snapshot
 
 log = structlog.get_logger()
